Fix Card edit and register for good
