+++ src/webapp/src/components/Library/index.js
@@ -5,6 +5,8 @@ import {
   Routes,
 } from 'react-router-dom';
 
+import Grid from '@mui/material/Grid';
+
 import LibraryLists from './lists';
 import SongList from './lists/albums/song-list';
 
@@ -12,22 +14,30 @@ const Library = () => {
   const lastListView = localStorage.getItem('libraryLastListView') || 'albums';
 
   return (
-    <Routes>
-      <Route
-        index
-        element={<Navigate to={lastListView} replace />}
-        exact
-      />
-      <Route
-        path="*"
-        element={<LibraryLists />}
-      />
-      <Route
-        path="albums/:artist/:album"
-        element={<SongList />}
-        exact
-      />
-    </Routes>
+    <Grid
+      container
+      id="library"
+      sx={{
+        padding: '10px',
+      }}
+    >
+      <Routes>
+        <Route
+          index
+          element={<Navigate to={lastListView} replace />}
+          exact
+        />
+        <Route
+          path="*"
+          element={<LibraryLists />}
+        />
+        <Route
+          path="albums/:artist/:album"
+          element={<SongList />}
+          exact
+        />
+      </Routes>
+    </Grid>
   );
 };
